+++ src/Crystalline.jl
@@ -105,7 +105,7 @@ include("double/groups.jl")
 include("double/irreps.jl")
 include("double/notation.jl")
 include("double/show.jl")
-export SU2, DSymOperation, su2, isbarred
+export SU2, DSymOperation, isbarred
 export DSpaceGroup, DPointGroup, DLittleGroup, DSiteGroup
 export doublegroup
 export DLGIrrep, DPGIrrep, DSiteIrrep